third_party/WebKit/LayoutTests/inspector-protocol/debugger/stepping-with-inline-source-map-and-blackboxing.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/inspector-protocol-test.js"></script>
<script type="text/javascript" src="resources/framework-with-inline-sourcemap.js"></script>
<script>
function boo()
{
    return 239;
}

function testFunction()
{
    debugger;
    foo(boo);
}

function test()
{
    var actions = [ "stepInto", "stepInto", "stepOut" ];

    InspectorTest.eventHandler["Debugger.paused"] = printStackTraceAndMakeNextStep;
    InspectorTest.sendCommand("Debugger.enable", {}, setSkipStackFrames);

    function setSkipStackFrames(response)
    {
        failIfError(response);
        InspectorTest.sendCommand("Debugger.skipStackFrames", { script: "foo\\.js$"}, callTestFunction);
    }

    function callTestFunction(response)
    {
        failIfError(response);
        InspectorTest.sendCommand("Runtime.evaluate", { "expression": "setTimeout(testFunction, 0)"});
    }

    function printStackTraceAndMakeNextStep(response)
    {
        failIfError(response);
        var callFrames = response.params.callFrames;
        InspectorTest.log("Stack trace:")
        for (var callFrame of callFrames) {
            var location = callFrame.functionLocation.lineNumber + ":" + callFrame.functionLocation.columnNumber;
            InspectorTest.log(callFrame.functionName + ":" + location);
        }

        var action = actions.shift();
        if (!action)
            InspectorTest.completeTest();
        else
            InspectorTest.sendCommand("Debugger." + action);
    }

    function failIfError(response)
    {
        if (response.error) {
            InspectorTest.log(JSON.stringify(response));
            InspectorTest.completeTest();
        }
    }
}
</script>
</head>
<body onload="runTest()">
</body>
</html>
